requirements.txt
# ── Streamlit ──────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Google APIs ────────────────────────────────────────────────────────────────
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth>=2.28.0          # includes service_account module

# ── Data processing ────────────────────────────────────────────────────────────
pandas>=2.1.0
openpyxl>=3.1.2              # optional: local .xlsx fallback

# ── LLM ────────────────────────────────────────────────────────────────────────
anthropic>=0.34.0

# ── Visualization (optional but recommended) ───────────────────────────────────
plotly>=5.18.0

# ── Config & utilities ─────────────────────────────────────────────────────────
PyYAML>=6.0.1
python-dotenv>=1.0.0         # local dev: load .env for API keys
